+++ ChangeLog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Had to bring back `GltfAsset.isDone` for render tests
+### Fixed
+-  Rendertests work again
+
 ## [3.0.1] - 2020-02-04
 ### Added
 - Error message when a UV set other than the first one is used (is unsupported; see issue #34)